pyproject.toml
[project]
name = "csv-handler-cli"
version = "0.1.0"
description = "Csv files handler, with filtering and aggregation functionality."
authors = [
    {name = "Petr Adzhigitov",email = "adzhigitovpetr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tabulate (>=0.9.0,<0.10.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[pytest]
pythonpath = "."

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.2.0"
pytest-cov = "^6.2.1"
allure-pytest = "^2.14.3"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "*Test"]
python_functions = "test_*"
